requirements.txt
toml>=0.10.2
google-genai>=0.2.0
python-dotenv>=1.0.0
baml-py==0.213.0
pydantic>=2.0.0
anthropic>=0.39.0
openai>=1.54.0
pytest>=8.0.0
